services/github-webhook-handler/api/requirements.txt
pydantic-settings
fastapi[standard]
structlog
opentelemetry-api
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-sdk